% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS fatal road incients by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The unique state number}

\item{year}{The year for which we'll be collecting data}
}
\value{
No value is returned. Will map the FARS data.
}
\description{
This function will collect FARS csv data for a given year and map the data for a given state number (state.num).
The code for this function is part of the Building R Packages course sourced from coursera.
Requires the dplyr, readr, maps, stats and graphics packages to be loaded.
Example data is only available for 2013:2015.
}
\examples{
\dontrun{
  setwd("data")
  year <- 2015
  state.num <- 15
  fars_map_state(state.num, year)
}

}
