% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv files into a data.frame / tbl_df}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to a csv file.}
}
\value{
Returns an object of the class data.frame, dbl_df, tbl.
}
\description{
This basic function takes a csv file path and reads the file using readr::read_csv and dplyr::tbl_df.
Note that dplyr::tbl_df (in the code) is deprecated and you should use tibble::as_tibble() instead.
The code for this function is part of the Building R Packages course sourced from coursera.
Example data is only available for 2013-2015.
Requires readr and dplyr to run.
}
\examples{
\dontrun{
  df <- fars_read(file.path("data", "accident_2013.csv.bz2"))
  print(class(df))
  head(df)
}

}
